feat: separate command & event handler
